backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
google-generativeai==0.3.2 